requirements-dev.txt
-r ./requirements.txt
decorator==4.0.4
pytest
flake8==2.5.0
webtest-aiohttp
invoke==0.11.1
